refactor: move main views into subpackage
