+++ src/main/resources/site/overwrite/auditranscribe/views/fxml/main/project-setup-view.fxml
@@ -13,7 +13,7 @@
 <?import javafx.scene.layout.VBox?>
 <?import javafx.scene.text.Font?>
 
-<AnchorPane fx:id="rootPane" prefHeight="400.0" prefWidth="500.0" xmlns="http://javafx.com/javafx/18" xmlns:fx="http://javafx.com/fxml/1" fx:controller="site.overwrite.auditranscribe.views.main.ProjectSetupViewController">
+<AnchorPane fx:id="rootPane" prefHeight="400.0" prefWidth="500.0" xmlns="http://javafx.com/javafx/18" xmlns:fx="http://javafx.com/fxml/1" fx:controller="site.overwrite.auditranscribe.views.main.view_controllers.ProjectSetupViewController">
    <stylesheets>
       <URL value="@../../css/base.css" />
       <URL value="@../../css/light-mode.css" />
